% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{calcular_diferencias}
\alias{calcular_diferencias}
\title{Title
Resultado de las diferencias entre elecciones, eligiendo una como contraste}
\usage{
calcular_diferencias(bd, partido, eleccion_referencia, eleccion_contraste)
}
\arguments{
\item{bd}{base de datos con resultados electorales}

\item{partido}{partidos o vector de partidos (ver base de datos de partidos)}

\item{eleccion_referencia}{elección con la que se contrastarán las elecciones.}

\item{eleccion_contraste}{elección o vector de elecciones contra los que se comparan
la elecciones de referencia.}
}
\value{
base con una columna de diferencias y  con el prefijo "dif_"
}
\description{
Title
Resultado de las diferencias entre elecciones, eligiendo una como contraste
}
\examples{
#NOT RUN
#calcular_diferencias(edomex::edomex_final, partido = c("pvem", "pt"),
#eleccion_referencia =  "df_21", eleccion_contraste = c("dl_21", "pm_21"))
}
