% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_censos.R
\name{calcular_irs}
\alias{calcular_irs}
\title{Calcula el índice de rezago para la elección y nivel deseado.}
\usage{
calcular_irs(bd, electoral, nivel, c_principal)
}
\arguments{
\item{bd}{Censo cargado para nivel en cuestión}

\item{electoral}{Base electoral cargada en el flujo de la clase Electoral}

\item{nivel}{nivel seleccionado: Sección, Municipio, Distrito local, Distrito Federal}

\item{c_principal}{color que se utilizará para representar el bajo rezago y permitirá obtener el complemento para el bajo rezago}
}
\value{
Una base con dos columnas adicionales: el cálculo del índice de rezago y el degradado de color asociado
}
\description{
Añade a la base electoral en cuestión y al nivel seleccionado en el momento el cálculo de rezago social.
}
