% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{calcular_relativos}
\alias{calcular_relativos}
\title{Calcula los votos relativos para los partidos requeridos}
\usage{
calcular_relativos(bd, partidos = bd$info$partidos, nominal)
}
\arguments{
\item{bd}{base larga con los partidos en una sola columna}

\item{partidos}{partidos de interés}

\item{nominal}{lista nominal de la unidad geográfica definida.}
}
\value{
base con dos columnas pct y label
}
\description{
Calcula los votos relativos para los partidos requeridos
}
