% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catalogo}
\alias{catalogo}
\title{Catálogo de bases de datos remotas}
\format{
A data frame with 275 rows and 3 variables:
\describe{
  \item{carpeta1}{Carpeta de primer nivel.}
  \item{capeta}{Subcarpeta.}
  \item{bd}{Nombre de la base de datos.}
}
}
\usage{
catalogo
}
\description{
Relación de carpetas y bases de datos disponibles en el almacenamiento remoto.
}
\keyword{datasets}
