% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claves}
\alias{claves}
\title{Claves de distritos electorales 2022}
\format{
A data frame with 242 rows and 5 variables:
\describe{
  \item{entidad}{Clave de la entidad.}
  \item{distritof_22}{Clave del distrito federal (2022).}
  \item{nombre_distritof_22}{Nombre del distrito federal.}
  \item{distritol_22}{Clave del distrito local (2022).}
  \item{nombre_distritol_22}{Nombre del distrito local.}
}
}
\usage{
claves
}
\description{
Claves y nombres de distritos federales y locales por entidad.
}
\keyword{datasets}
