% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claves_mun}
\alias{claves_mun}
\title{Claves de municipios 2022}
\format{
A data frame with 2477 rows and 2 variables:
\describe{
  \item{municipio_22}{Clave del municipio (2022).}
  \item{nombre_municipio_22}{Nombre del municipio.}
}
}
\usage{
claves_mun
}
\description{
Claves y nombres de municipios.
}
\keyword{datasets}
