% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{colorear_indice}
\alias{colorear_indice}
\title{Creación de paletas para índice}
\usage{
colorear_indice(bd, c_principal, var)
}
\arguments{
\item{bd}{Base de datos con columna de índice.}

\item{c_principal}{el color que tomará el valor más alto del índice}

\item{var}{la variable de tipo índice a la que se le asociará el color.}
}
\description{
Añade una nueva columna a la base de datos. Recibe una variable de tipo índice, calcula el color complementario
a un color entregado y relaciona los valores del índice con un color de la paleta.
}
