% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{crear_indice}
\alias{crear_indice}
\title{Cálculo de índice para variables electorales}
\usage{
crear_indice(bd, partido, nivel)
}
\arguments{
\item{bd}{Base de datos en términos relativos}

\item{partido}{variable para la cual se quiere calcular el índice.}

\item{nivel}{El nivel de agregación en el que se encuentran los datos. Si la base de datos tiene más de uno, se pueden incluir dentro de la función.}
}
\value{
base con columnas añadidas de cada partido para el cual se calculó el índice
}
\description{
Función que calcula el índice para la variable o variables que se le soliciten.
Requiere que la base de datos esté en términos relativos.
La función estandariza los resultados para que no importe el signo de los datos.
}
