% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{crear_label}
\alias{crear_label}
\title{Crea la label para el leaflet con todas las variables que se suelen considerar relevantes en un flujo.}
\usage{
crear_label(bd, nivel)
}
\arguments{
\item{bd}{Base que contiene todas las variables relevantes}

\item{nivel}{Alguna de las siguientes opciones: seccion, municipio, distrito local o distrito federal}
}
\value{
la base original con una columna adicional con la label.
}
\description{
Crea la label para el leaflet con todas las variables que se suelen considerar relevantes en un flujo.
}
