% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{crear_mapa}
\alias{crear_mapa}
\title{Mapa electoral estático}
\usage{
crear_mapa(shp, fill, linewidth)
}
\arguments{
\item{shp}{shapefile que se quiere dibujar, tiene que tener una variable con el string de los colores que se van a mostrar.}

\item{fill}{variable que se quiere colorear en el mapa. Para índice de morenismo se usa 'morena', para resultados de una elección se usa la elección 'pm_21'}

\item{linewidth}{el ancho de las líneas del mapa}
}
\value{
objeto de tipo ggplot
}
\description{
Mapa electoral estático
}
