% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{crear_mapa_electoral}
\alias{crear_mapa_electoral}
\title{Title
Se utiliza un método matemático llamada análisis de componentes principales (PCA por sus siglas en inglés) que captura los patrones de votación y los sintetiza en una gráfica de dos dimensiones.}
\usage{
crear_mapa_electoral(bd, eleccion, anio, grupo = NA, partido = NA)
}
\arguments{
\item{bd}{base de datos con resultados electorales}

\item{eleccion}{elección o vector de elecciones de la cual se van a obtener los ganadores}

\item{anio}{año de las elecciones seleccionadas}

\item{grupo}{nivel de observación de la gráfica (secciones, municipios, distritos)}

\item{partido}{vector que partidos que se van a graficar}
}
\value{
gráfica de modelo pca
}
\description{
Title
Se utiliza un método matemático llamada análisis de componentes principales (PCA por sus siglas en inglés) que captura los patrones de votación y los sintetiza en una gráfica de dos dimensiones.
}
