% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{crear_quantiles}
\alias{crear_quantiles}
\title{Divide los datos en 4 cuantiles}
\usage{
crear_quantiles(bd, partido, grupos = 4)
}
\arguments{
\item{bd}{Base con la variable que se quiere clasificar}

\item{partido}{partido para el cual se quiere realizar la clasificación}

\item{grupos}{Número de grupos en los que se quiere clasificar los datos. 4 es el valor default.}
}
\value{
Un base de datos con una columna adicional, cuyo nombre comenzará con 'quant'
}
\description{
Función auxiliar a `crear_indice`, divide el índice obtenido en 4 grupos y lo convierte en factor.
}
