% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiones.R
\name{d0_d1}
\alias{d0_d1}
\title{Function Name: d0_d1}
\usage{
d0_d1(bd, shp, K, tipo = "gower")
}
\arguments{
\item{bd}{Data frame containing the data to be clustered. The function expects this data frame to contain columns starting with "ele_", which are transformed using a logarithm.}

\item{shp}{A shapefile (spatial data) used for calculating spatial distances.}

\item{K}{An integer representing the number of clusters to be used in the ClustGeo algorithm.}

\item{tipo}{A string specifying the type of distance metric to be used. "gower" for Gower's distance (default) and any other value for Euclidean distance.}
}
\value{
Returns a list containing the following elements:
  - d0: A distance matrix (either Gower's or Euclidean) computed from the 'bd' data.
  - d1: A distance matrix computed from the 'shp' shapefile.
  - cr: Clustering results obtained from the ClustGeo::choicealpha function.
  - shp: The original shapefile passed to the function.
}
\description{
This function performs spatial and attribute clustering analysis using either Gower's distance or Euclidean distance. It takes a dataset, a shapefile, a number 'K' for clustering, and a type of distance metric (defaulting to "gower"). It returns a list containing distance matrices and clustering results along with the original shapefile.
}
