% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_colores.R
\name{degradar_color_partido}
\alias{degradar_color_partido}
\title{Title}
\usage{
degradar_color_partido(
  bd_larga,
  nombre,
  variable,
  colores_nombrados,
  valor_maximo = 1
)
}
\arguments{
\item{bd_larga}{base electoral en long}

\item{nombre}{partido de interés}

\item{variable}{unidad de interés que se desea analizar (votos, pocentaje, etc)}

\item{colores_nombrados}{vector compuesto con los nombres de partidos y colores que le corresponden}

\item{valor_maximo}{valor máximo que toma la generación de degradados}
}
\value{

}
\description{
Title
}
