% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diccionario}
\alias{diccionario}
\title{Diccionario de entidades}
\format{
A data frame with 32 rows and 3 variables:
\describe{
  \item{id_estado}{Clave numérica del estado.}
  \item{estado}{Nombre del estado.}
  \item{abreviatura}{Abreviatura del estado.}
}
}
\usage{
diccionario
}
\description{
Catálogo de entidades federativas con id, nombre y abreviatura.
}
\keyword{datasets}
