% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diccionario_censo20}
\alias{diccionario_censo20}
\title{Diccionario del censo}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{indicador}{descripcion de la variable}
  \item{definicion}{Rango que abarca la variable}
  \item{variable}{abreviatura de la variable tal como aparece en el censo}
  \item{unidad_de_medida}{unidad de medida}
  \item{unidad_de_observacion}{unidad de observacion}
  \item{poblacion_de_referencia}{población de referencia}
  \item{denominador}{denominador que relativiza la variable}
  \item{tema}{tema}
}
}
\usage{
diccionario_censo20
}
\description{
Diccionario del censo con información para le calculo relativo de variables
}
\keyword{datasets}
