% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{ejecutar_sankey}
\alias{ejecutar_sankey}
\title{Gráfica sankey}
\usage{
ejecutar_sankey(bd, colores)
}
\arguments{
\item{bd}{base procesada mediante `procesar_sankey`}

\item{colores}{colores que se utilizan para el sankey}
}
\value{
objeto tipo ggplot
}
\description{
Gráfica sankey
}
