% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterios.R
\name{eliminar_especiales}
\alias{eliminar_especiales}
\title{Elimina las casillas especiales (tipo_casilla = "S")}
\usage{
eliminar_especiales(bd)
}
\arguments{
\item{bd}{Base de datos a la que se le eliminan las casilla especiales}
}
\value{
Data frame sin casillas especiales (tipo_casilla = "S")
}
\description{
Elimina las casillas especiales (tipo_casilla = "S")
}
\examples{
 eliminar_especiales(bd)
}
