% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterios.R
\name{eliminar_votoExtranjero}
\alias{eliminar_votoExtranjero}
\title{Se eliminan los votos que están en las secciones = 0000}
\usage{
eliminar_votoExtranjero(bd)
}
\arguments{
\item{bd}{Base de datos a la que se le eliminan los votos en el extranjero}
}
\value{
Data frame sin las secciones iguales a 0000
}
\description{
Se eliminan los votos que están en las secciones = 0000
}
\examples{
bd \%>\%  eliminar_votoExtranjero()
}
