% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentabilidad.R
\name{extraer_anios_eleccion}
\alias{extraer_anios_eleccion}
\title{Extrae los años disponibles para un tipo de elección}
\usage{
extraer_anios_eleccion(obj_eleccion, eleccion)
}
\arguments{
\item{obj_eleccion}{Objeto `Tablero` o `Electoral` con `$info$elecciones_agregadas`.}

\item{eleccion}{Tipo de elección: `"pm"`, `"dl"`, `"df"`, `"pr"`, `"sen"`, `"gb"`.}
}
\value{
Vector numérico de años ordenado de menor a mayor.
}
\description{
Extrae los años disponibles para un tipo de elección
}
