% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{extraer_partidos}
\alias{extraer_partidos}
\title{Extrae los partidos en una base}
\usage{
extraer_partidos(bd, eleccion, tipo)
}
\arguments{
\item{bd}{es la base a evaluar}

\item{eleccion}{elección seleccionada}

\item{tipo}{si los datos están en términos relativos o absolutos}
}
\value{
vector con los partidos seleccionados
}
\description{
Extrae los partidos en una base
}
