% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{ganador}
\alias{ganador}
\title{Función para señalae quién fue el ganador por eleccion por nivel}
\usage{
ganador(bd, nivel, eleccion)
}
\arguments{
\item{bd}{Base de datos electoral de la que se quiere sacar el ganador}

\item{nivel}{Nivel en el que se determinan las alianzas dependiendo de la unidad en la que se realiza la elección.}

\item{eleccion}{Es el tipo de elección y su año separado por "_". Opciones posibles para 2021: pm_21, dl_21, df_21.}
}
\value{
tibble con la columna de ganador con los ganadores por nivel
}
\description{
Función para señalae quién fue el ganador por eleccion por nivel
}
