% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{ganador_eleccion}
\alias{ganador_eleccion}
\title{Title
Obtiene el partido ganador para cada una de las observaciones de la base. Puede ser sección, municipio o distrito}
\usage{
ganador_eleccion(bd, eleccion, tipo = NULL, nivel, partido = NULL)
}
\arguments{
\item{bd}{base de datos con resultados electorales}

\item{eleccion}{elección o vector de elecciones de la cual se van a obtener los ganadores}

\item{tipo}{tipo de resultado a usar (`"relativo"` o absoluto). Por default `NULL`.}

\item{nivel}{nivel de observación (sección, municipio o distrito).}

\item{partido}{partido de interés para filtrar el ganador. Por default `NULL`.}
}
\value{
base de datos con una columna que indica el partido ganador de cada eleccion referida
}
\description{
Title
Obtiene el partido ganador para cada una de las observaciones de la base. Puede ser sección, municipio o distrito
}
