% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_data.R
\name{get_aelectoral_cache}
\alias{get_aelectoral_cache}
\title{Get the local cache directory for aelectoral2}
\usage{
get_aelectoral_cache()
}
\value{
Character string. The path to the local cache directory.
}
\description{
Get the local cache directory for aelectoral2
}
\keyword{internal}
