% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{graficar_barras}
\alias{graficar_barras}
\title{Configuración general para gráfica de barras}
\usage{
graficar_barras(bd, x, y, fill, label, colores, eje_x, eje_y, size = 12)
}
\arguments{
\item{bd}{base electoral}

\item{x}{variable que irá en el eje x de la gráfica}

\item{y}{variable que irá en el eje y de la gráfica}

\item{fill}{variable por la que se colorea}

\item{label}{etiquetas de las barras}

\item{colores}{los colores a los que se asocia el parámetro fill - se utiliza scales_fill_manual}

\item{eje_x}{etiqueta del eje x}

\item{eje_y}{etiqueta del eje y}

\item{size}{tamaño de la letra de gráfico, parámetro de 'base_size'}
}
\value{
objeto de ggplot
}
\description{
Configuración general para gráfica de barras
}
