% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{graficar_cloropeta}
\alias{graficar_cloropeta}
\title{Title
grafica una cloropeta con la distribución del apoyo del partido solicitado.
una vez coloreada la sección se hace un degradado en función de la intensidad con la que dicho partido resultó ganador.
Es decir, se colorea más fuerte mientras mayor haya sido la victoria y menos fuerte si fue una sección muy competida.}
\usage{
graficar_cloropeta(bd, shp, colores_nombrados, eleccion, grupo)
}
\arguments{
\item{bd}{base de datos con resultados electorales}

\item{shp}{archivo .shp con el polígono del nivel de observación correspondiente}

\item{colores_nombrados}{vector compuesto con los nombres de partidos y colores que le corresponden}

\item{eleccion}{elección o vector de elecciones de la cual se van a obtener los ganadores}

\item{grupo}{nivel de observación de la gráfica (secciones, municipios, distritos)}
}
\value{
mapa con la intensidad de apoyo por partido
}
\description{
Title
grafica una cloropeta con la distribución del apoyo del partido solicitado.
una vez coloreada la sección se hace un degradado en función de la intensidad con la que dicho partido resultó ganador.
Es decir, se colorea más fuerte mientras mayor haya sido la victoria y menos fuerte si fue una sección muy competida.
}
