% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{graficar_pointrange}
\alias{graficar_pointrange}
\title{Gráfica tipo pointrange}
\usage{
graficar_pointrange(bd, eje_x, grupo, colores, indice, size = 12)
}
\arguments{
\item{bd}{base producto de `procesar_pointrange`}

\item{eje_x}{variable del eje x}

\item{grupo}{variable que puede tomar dos valores 'partidos' o 'elecciones'}

\item{colores}{colores que se utilizan dentro de un scale_color_manual}

\item{indice}{indice graficado}

\item{size}{tamaño del base_size}
}
\value{
objeto tipo ggplot
}
\description{
Gráfica tipo pointrange
}
