% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{graficar_tiles}
\alias{graficar_tiles}
\title{Grafica un geom_tiles}
\usage{
graficar_tiles(bd, x, y, low, high, name, eje_x, eje_y, size = 12)
}
\arguments{
\item{bd}{base electoral con variable 'quant_'}

\item{x}{variable categórica que se mostrará en el eje x}

\item{y}{variable categórica que se mostrará en el eje y}

\item{low}{elemento más chico del gradiente}

\item{high}{elemento más grande del gradiente}

\item{name}{nombre que se colocará en la leyenda}

\item{eje_x}{título del eje x}

\item{eje_y}{título del eje y}

\item{size}{tamaño del base_size}
}
\value{
objeto de ggplot
}
\description{
Compara la coincidencia de dos variables categóricas
}
