% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{graficar_totales_eleccion}
\alias{graficar_totales_eleccion}
\title{Title
Gráfica de líneas con el total de votos comparándolos entre elecciones.}
\usage{
graficar_totales_eleccion(bd, colores_nombrados, eleccion, grupo = NULL)
}
\arguments{
\item{bd}{base de datos con resultados electorales}

\item{colores_nombrados}{vector compuesto con los nombres de partidos y colores que le corresponden}

\item{eleccion}{elección o vector de elecciones de la cual se van a obtener los ganadores}

\item{grupo}{nivel de observación de la gráfica (secciones, municipios, distritos)}
}
\value{
Grafica de líneas con los totales de las elecciones solicitadas
}
\description{
Title
Gráfica de líneas con el total de votos comparándolos entre elecciones.
}
