% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{graficar_violin}
\alias{graficar_violin}
\title{Gráfica de distribución tipo violín}
\usage{
graficar_violin(bd, x, y, fill, colores, eje_x, eje_y, size = 12)
}
\arguments{
\item{bd}{base electoral}

\item{x}{variable que irá en el eje x de la gráfica}

\item{y}{variable que irá en el eje y de la gráfica}

\item{fill}{variable por la que se colorea}

\item{colores}{los colores a los que se asocia el parámetro fill - se utiliza scales_fill_identity}

\item{eje_x}{etiqueta del eje x}

\item{eje_y}{etiqueta del eje y}

\item{size}{tamaño de la letra de gráfico, parámetro de 'base_size', por default tiene tamaño 12}
}
\value{
objeto tipo ggplot
}
\description{
Gráfica de distribución tipo violín
}
