% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_base_electoral.R
\name{insertar_sufijo}
\alias{insertar_sufijo}
\title{Title}
\usage{
insertar_sufijo(bd, eleccion_nivel, eleccion_year)
}
\arguments{
\item{bd}{base de datos ya procesada}

\item{eleccion_nivel}{nivel de las elección a la que se le pone el sufijo (df,dl,pm)}

\item{eleccion_year}{año de la eleción de interés (15,18,21)}
}
\value{

}
\description{
Title
}
