% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura.R
\name{join_shp_bd}
\alias{join_shp_bd}
\title{Para juntar un shapefile con otra base de datos
Funcion basada en left_join}
\usage{
join_shp_bd(shp, bd)
}
\arguments{
\item{shp}{Base de datos de tipo shp}

\item{bd}{Base de datos que se va a unir con el shapefile}
}
\value{
Un shp unido con bd
}
\description{
Para juntar un shapefile con otra base de datos
Funcion basada en left_join
}
\examples{
join_shp_bd(secc_21, df_21)
}
