% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura.R
\name{leer_alianza}
\alias{leer_alianza}
\title{Función para leer la base de datos de alianzas por partido y por nivel
También transforma la bd de alianzas al nivel de la base electoral.}
\usage{
leer_alianza(nivel, eleccion, entidad, bd_e)
}
\arguments{
\item{nivel}{Nivel en el que se determinan las alianzas dependiendo de la unidad en la que se realiza la elección.}

\item{eleccion}{Es el tipo de elección y su año separado por "_". Opciones posibles para 2021: pm_21, dl_21, df_21.}

\item{entidad}{Cuando es nacional es "nac", cuando es local se pone la abreviatura oficial, por ejemplo "chis", "dgo", "mex".}

\item{bd_e}{Base de datos electoral a la ue se le van a pegar las coaliciones por partido.}
}
\value{
Regresa un data frame de alianzas
}
\description{
Función para leer la base de datos de alianzas por partido y por nivel
También transforma la bd de alianzas al nivel de la base electoral.
}
\examples{
leer_alianza(nivel, eleccion, self$entidad, self$bd)
}
