% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura.R
\name{leer_censo}
\alias{leer_censo}
\title{Función para leer censo
Basada en la funcion read_rds del paquete readr}
\usage{
leer_censo(ano, entidad, nivel)
}
\arguments{
\item{ano}{Es el año del censo}

\item{entidad}{Cuando se requiere la información de todo el país se escribe "nacional", cuando es local se pone la abreviatura oficial de la entidad, por ejemplo "chis", "dgo", "mex".}

\item{nivel}{Unidad geográfica en la que se requiere la información del censo}
}
\value{
tibble de la base del censo
}
\description{
Función para leer censo
Basada en la funcion read_rds del paquete readr
}
\examples{
leer_censo(ano = 2020, entidad = "nacional", nivel = "seccion")
}
