% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura.R
\name{leer_shp}
\alias{leer_shp}
\title{Lee un shapefile}
\usage{
leer_shp(unidad, entidad)
}
\arguments{
\item{unidad}{si es de municipio, estado, distrito, seccion, etc.}

\item{entidad}{el estado de donde es}
}
\value{
shp
}
\description{
Lee un shapefile
}
\examples{
leer_shp(unidad, entidad)
}
