% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiones.R
\name{mapa_region}
\alias{mapa_region}
\title{Dibuja el mapa de una región}
\usage{
mapa_region(d0_d1, alpha)
}
\arguments{
\item{d0_d1}{base obtenida de la función `d0_d1`}

\item{alpha}{transparencia del leaflet}
}
\value{
un leaflet
}
\description{
Dibuja el mapa de una región
}
