% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{nominal}
\alias{nominal}
\title{Obtienes la lista nominal total de la base y la elección seleccionada}
\usage{
nominal(bd, eleccion)
}
\arguments{
\item{bd}{base de datos con columna nominal}

\item{eleccion}{elección de interés}
}
\value{
vector numérico
}
\description{
Obtienes la lista nominal total de la base y la elección seleccionada
}
