% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{obtener_absolutos}
\alias{obtener_absolutos}
\title{Extrae los datos absolutos de una base de datos electoral}
\usage{
obtener_absolutos(bd, eleccion)
}
\arguments{
\item{bd}{Base de datos electoral}

\item{eleccion}{elecciones de interés ('gb_17')}
}
\value{
subconjunto de valores en términos absolutos para la base entrante
}
\description{
Extrae los datos absolutos de una base de datos electoral
}
