% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paleta}
\alias{paleta}
\title{Paleta de colores de partidos}
\format{
A data frame with 27 rows and 3 variables:
\describe{
  \item{partidos}{Clave del partido.}
  \item{colores}{Color hexadecimal.}
  \item{alcance}{Alcance del partido (\code{"nacional"} o entidad).}
}
}
\usage{
paleta
}
\description{
Colores por partido y su alcance (nacional o estatal).
}
\keyword{datasets}
