% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_explorar.R
\name{probar_independencia_ganador}
\alias{probar_independencia_ganador}
\title{Title}
\usage{
probar_independencia_ganador(bd, ganador, eleccion, ...)
}
\arguments{
\item{bd}{base con resultados electorales y una columna adicional con el ganador de cada sección}

\item{ganador}{partido del que se busca analizar la independencia de su resultado}

\item{eleccion}{elección elegida para analizar}

\item{...}{variables del censo de las que se busca analizar la independencia}
}
\value{

}
\description{
Title
}
