% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{procesar_pointrange}
\alias{procesar_pointrange}
\title{Procesamiento para gráfica pointrange}
\usage{
procesar_pointrange(bd, indice, partidos, elecciones = NULL)
}
\arguments{
\item{bd}{base de datos electoral con columna 'quant_'}

\item{indice}{los cuantiles que se van a graficar en el eje x}

\item{partidos}{los partidos para los que se calcula el pointrange}

\item{elecciones}{en caso de que no sean partidos, sino elecciones, por default es NULL}
}
\value{
base con datos procesado para gráfica pointrange
}
\description{
Procesamiento para gráfica pointrange
}
