% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{procesar_sankey}
\alias{procesar_sankey}
\title{Procesamiento para Sankey}
\usage{
procesar_sankey(bd, elecciones)
}
\arguments{
\item{bd}{base desagregada a nivel sección con columna de ganador}

\item{elecciones}{Elecciones de interés para visualización de sankey}
}
\value{
base en formato idóneo para sankey
}
\description{
Procesamiento para Sankey
}
