% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizaciones.R
\name{procesar_secciones_ganadas}
\alias{procesar_secciones_ganadas}
\title{Calcula el número de secciones ganadas por ganador}
\usage{
procesar_secciones_ganadas(bd, eleccion)
}
\arguments{
\item{bd}{base con columna de ganadores}

\item{eleccion}{elección de interes}
}
\value{
base procesada con estadísticos de secciones ganadas por partido
}
\description{
Calcula el número de secciones ganadas por ganador
}
