% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura.R
\name{reducir}
\alias{reducir}
\title{Base de datos que resume agrupando por las llaves
Basada en la función summarise}
\usage{
reducir(bd, completa, llaves)
}
\arguments{
\item{bd}{Base de datos que se quiere reducir}

\item{completa}{base de datos electoral}

\item{llaves}{Son las claves cartográficas de los niveles. Por default la unidad mínima es sección y está acompañada de estado.}
}
\value{
Data frame
}
\description{
Base de datos que resume agrupando por las llaves
Basada en la función summarise
}
\examples{
add \%>\% reducir(self$bd, self$llaves)
}
