% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relacion_ine_inegi}
\alias{relacion_ine_inegi}
\title{Relación INE - INEGI}
\format{
A data frame with 302 rows and 3 variables:
\describe{
  \item{entidad}{Clave de la entidad.}
  \item{mun}{Clave municipal (INE).}
  \item{municipio_22}{Clave municipal (INEGI, 2022).}
}
}
\usage{
relacion_ine_inegi
}
\description{
Equivalencia entre claves municipales del INE y del INEGI.
}
\keyword{datasets}
