% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_censos.R
\name{relativizar_censo}
\alias{relativizar_censo}
\title{Relativiza el censo}
\usage{
relativizar_censo(bd)
}
\arguments{
\item{bd}{Base de datos del censo a relativizar.}
}
\description{
Función de poca utilidad, se valorará su remoción
}
