% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{repartir_candidato}
\alias{repartir_candidato}
\title{Reparte los votos por candidato tomando en cuenta si los candidatos fueron con alianzas o en candidatura común con más de un partido o no.}
\usage{
repartir_candidato(bd, al, nivel, eleccion)
}
\arguments{
\item{bd}{base de datos electoral que se quiere repartir por candidatos (tomando en cuenta alianzas)}

\item{al}{alianzas}

\item{nivel}{Nivel en el que se determinan las alianzas dependiendo de la unidad en la que se realiza la elección.}

\item{eleccion}{Es el tipo de elección y su año separado por "_". Opciones posibles para 2021: pm_21, dl_21, df_21.}
}
\value{
Base de datos repartida por candidato
}
\description{
Reparte los votos por candidato tomando en cuenta si los candidatos fueron con alianzas o en candidatura común con más de un partido o no.
}
\examples{
repartir_candidato(bd = self$bd_partido[[eleccion]], alianzas, nivel, eleccion)
}
