% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partidos.R
\name{repartir_coalicion}
\alias{repartir_coalicion}
\title{Para repartir votos por coalición}
\usage{
repartir_coalicion(bd, nivel, eleccion)
}
\arguments{
\item{bd}{Base de datos que se quiere repartir}

\item{nivel}{Nivel en el que se determinan las alianzas dependiendo de la unidad en la que se realiza la elección.}

\item{eleccion}{Es el tipo de elección y su año separado por "_". Opciones posibles para 2021: pm_21, dl_21, df_21.}
}
\value{
Base de datos repartida
}
\description{
Para repartir votos por coalición
}
\examples{
elf$bd \%>\% repartir_coalicion(nivel = nivel, eleccion = eleccion)

}
