% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterios.R
\name{repartir_especiales}
\alias{repartir_especiales}
\title{Reparte votos en tipo_casilla = "S"}
\usage{
repartir_especiales(bd)
}
\arguments{
\item{bd}{Base de datos a repartir}
}
\value{
Data frame con las casillas especiales (tipo_casilla = "S") repartidas
}
\description{
Reparte votos en tipo_casilla = "S"
}
\examples{
repartir_especiales(bd)
}
