% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{t_nac}
\alias{t_nac}
\title{Te trae el resultado de votos de una variable seleccionada con un patron de una bd electoral.}
\usage{
t_nac(bd, tipo, patron)
}
\arguments{
\item{bd}{Base de datos a testear}

\item{tipo}{si es por partido o por candidato}

\item{patron}{la variable de la que se quiere el resultado}
}
\value{
El resultado electoral con el método señalado
}
\description{
Te trae el resultado de votos de una variable seleccionada con un patron de una bd electoral.
}
\examples{
aux <- mex$bd \%>\% repartir_coalicion(nivel = "distritof_21", eleccion = "df_21")
t_nac(aux, "partido", "morena_")
}
