% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{t_nivel}
\alias{t_nivel}
\title{Devuelve el resultado electoral por partido o por candidato de una variable seleccionada con un patron de un estado en particular en un nivel seleccionado}
\usage{
t_nivel(bd, tipo, patron, estado, nivel, seleccion)
}
\arguments{
\item{bd}{Base de datos a testear}

\item{tipo}{si es por partido o por candidato}

\item{patron}{la variable de la que se quiere el resultado}

\item{estado}{Número de la entidad del ine}

\item{nivel}{Nivel en el que se determinan las alianzas dependiendo de la unidad en la que se realiza la elección.}

\item{seleccion}{que selección de acuerdo con el nivel se requiere}
}
\value{
Resultado electoral por nivel
}
\description{
Devuelve el resultado electoral por partido o por candidato de una variable seleccionada con un patron de un estado en particular en un nivel seleccionado
}
\examples{
mex$bd \%>\%  t_nivel("candidato", "ele_morena_dl_21", estado = 15, nivel = "estado",  seleccion = "15")
}
