% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura.R
\name{anadir_cc}
\alias{anadir_cc}
\title{Para incluir las candidaturas comunes en el análisis}
\usage{
anadir_cc(bd, eleccion)
}
\value{
el vector de self$partido con las candidaturas comunes
}
\description{
Para incluir las candidaturas comunes en el análisis
}
