% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_colores.R
\name{asociar_colores}
\alias{asociar_colores}
\title{Asigna a un objeto de la clase los colores de los partidos seleccionados}
\usage{
asociar_colores(partidos)
}
\arguments{
\item{partidos}{}
}
\value{
Un vector nombrado con los colores de los partidos seleccionados
}
\description{
Asigna a un objeto de la clase los colores de los partidos seleccionados
}
