% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_colores.R
\name{colorear_ganador_degradado}
\alias{colorear_ganador_degradado}
\title{Obtener degradado de variable seleccionada}
\usage{
colorear_ganador_degradado(
  bd,
  eleccion,
  colores_nombrados,
  grupo,
  tipo,
  saturacion = 0.9
)
}
\arguments{
\item{bd}{base con resultados electorales y una columna adicional con el ganador de cada sección}

\item{eleccion}{elección elegida para analizar}

\item{colores_nombrados}{vector compuesto con los nombres de partidos y colores que le corresponden}

\item{grupo}{nivel de observación de la gráfica (secciones, municipios, distritos)}

\item{saturacion}{nivel de saturación de la paleta de color}
}
\value{
base con columnas adicionales
}
\description{
Obtener degradado de variable seleccionada
}
