% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_base_electoral.R
\name{detectar_partidos}
\alias{detectar_partidos}
\title{#' Title
#'
#' @param bd base de datos de cualquier elección en crudo (sin ningún procesamiento)
#' @param identificadores nivel de observación para el que se aplica el análisis
#' @param partido los partidos identificados para cada elección. Calculados con la función detectar partidos
#'
#' @return
#' @export
#'
#' @examples}
\usage{
detectar_partidos(bd)
}
\arguments{
\item{bd}{base de datos de cualquier elección en crudo (sin ningún procesamiento)}
}
\value{

}
\description{

}
\details{

}
