% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.R
\name{Electoral}
\alias{Electoral}
\title{Title}
\value{

}
\description{
Title

Title
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Electoral$new()}}
\item \href{#method-obtener_bd}{\code{Electoral$obtener_bd()}}
\item \href{#method-clone}{\code{Electoral$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electoral$new(
 
    inicial = "~/Dropbox (Selva)/Ciencia de datos/Consultoría Estadística/Recursos/Externos/INE/Bases de datos/Resultados definitivos",
  ano,
  tipo,
  eleccion,
  entidad,
  normal = T,
  nivel = "seccion"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nivel}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obtener_bd"></a>}}
\if{latex}{\out{\hypertarget{method-obtener_bd}{}}}
\subsection{Method \code{obtener_bd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electoral$obtener_bd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Electoral$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
