% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.R
\name{ElectoralSHP}
\alias{ElectoralSHP}
\title{Clase R6 para leer y unir shapefiles}
\description{
Se inicia leyendo un shp
}
\details{
Al shp leído se le pude agregar otrabase de datos
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shp}}{lista con los shps cargados}

\item{\code{entidades}}{entidad para la cual se quiere cargar el shp}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElectoralSHP-new}{\code{ElectoralSHP$new()}}
\item \href{#method-ElectoralSHP-print}{\code{ElectoralSHP$print()}}
\item \href{#method-ElectoralSHP-agregar_shp}{\code{ElectoralSHP$agregar_shp()}}
\item \href{#method-ElectoralSHP-juntar_bd}{\code{ElectoralSHP$juntar_bd()}}
\item \href{#method-ElectoralSHP-clone}{\code{ElectoralSHP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectoralSHP-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElectoralSHP-new}{}}}
\subsection{Method \code{new()}}{
Lee un shapefile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$new(unidad, entidad)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unidad}}{si es de municipio, estado, distrito, seccion, etc.}

\item{\code{entidad}}{el estado de donde es}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Una lista con shapefiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectoralSHP-print"></a>}}
\if{latex}{\out{\hypertarget{method-ElectoralSHP-print}{}}}
\subsection{Method \code{print()}}{
muestra qué se ha incluído en la clase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectoralSHP-agregar_shp"></a>}}
\if{latex}{\out{\hypertarget{method-ElectoralSHP-agregar_shp}{}}}
\subsection{Method \code{agregar_shp()}}{
añade otro shp al objeto shp
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$agregar_shp(unidad, entidad = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unidad}}{nivel de shp que se quiere cargar}

\item{\code{entidad}}{entidad para la que se quiere datos, se hereda del método `Electoral$new`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
lista con un nuevo shp
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectoralSHP-juntar_bd"></a>}}
\if{latex}{\out{\hypertarget{method-ElectoralSHP-juntar_bd}{}}}
\subsection{Method \code{juntar_bd()}}{
Junta shapefiles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$juntar_bd(nivel, bd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nivel}}{Si la base que se va a juntar es por seccion, municipio, distrito, etc}

\item{\code{bd}}{base de datos que se le quiere pegar al shp}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElectoralSHP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElectoralSHP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
