% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.R
\name{ElectoralSHP}
\alias{ElectoralSHP}
\title{Clase R6 para leer y unir shapefiles}
\description{
Se inicia leyendo un shp
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ElectoralSHP$new()}}
\item \href{#method-print}{\code{ElectoralSHP$print()}}
\item \href{#method-agregar_shp}{\code{ElectoralSHP$agregar_shp()}}
\item \href{#method-juntar_bd}{\code{ElectoralSHP$juntar_bd()}}
\item \href{#method-clone}{\code{ElectoralSHP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Lee un shapefile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$new(unidad, entidad)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unidad}}{si es de municipio, estado, distrito, seccion, etc.}

\item{\code{entidad}}{el estado de donde es}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Una lista con shapefiles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agregar_shp"></a>}}
\if{latex}{\out{\hypertarget{method-agregar_shp}{}}}
\subsection{Method \code{agregar_shp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$agregar_shp(unidad, entidad = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-juntar_bd"></a>}}
\if{latex}{\out{\hypertarget{method-juntar_bd}{}}}
\subsection{Method \code{juntar_bd()}}{
Junta shapefiles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$juntar_bd(nivel, bd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nivel}}{Si la base que se va a juntar es por seccion, municipio, distrito, etc}

\item{\code{bd}}{base de datos que se le quiere pegar al shp}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElectoralSHP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
