% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.R
\name{Graficas}
\alias{Graficas}
\title{Clase R6 para procesar y graficar los datos electorales}
\description{
Se inicia definiendo la unidad geográfica de trabajo mediante el método $filtrar.
Se puede definir un nivel y una unidad geográfica.
}
\details{
Algunas funciones tienen parámetros espécificos, pero otras solo tienes que llamarlas y usan los datos ya incluidos en la clase tablero.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tab}}{Recibe como un único insumo la clase tablero
Initialize: Obtener los datos incorporados en clase}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Graficas-new}{\code{Graficas$new()}}
\item \href{#method-Graficas-mapa}{\code{Graficas$mapa()}}
\item \href{#method-Graficas-secciones_ganadas}{\code{Graficas$secciones_ganadas()}}
\item \href{#method-Graficas-voto_relativo}{\code{Graficas$voto_relativo()}}
\item \href{#method-Graficas-distribucion_participacion}{\code{Graficas$distribucion_participacion()}}
\item \href{#method-Graficas-sankey}{\code{Graficas$sankey()}}
\item \href{#method-Graficas-pointrange}{\code{Graficas$pointrange()}}
\item \href{#method-Graficas-tiles}{\code{Graficas$tiles()}}
\item \href{#method-Graficas-clone}{\code{Graficas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-new"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-new}{}}}
\subsection{Method \code{new()}}{
Recupera los datos de la clase tablero
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$new(tablero)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tablero}}{Clase tablero}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-mapa"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-mapa}{}}}
\subsection{Method \code{mapa()}}{
crea un mapa con el coloreo que se desee
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$mapa(seccion, fill, linewidth = 0.6, labels = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seccion}}{variable booleana para especificar que el mapa que se quiere ese de secciones o no}

\item{\code{fill}}{variable con la cual se va a colorear el mapa}

\item{\code{linewidth}}{ancho de la línea, parámetro heredado de `geom_sf`}

\item{\code{labels}}{booleano para definir si se quieren leyendas en el mapa}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
objeto tipo ggplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-secciones_ganadas"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-secciones_ganadas}{}}}
\subsection{Method \code{secciones_ganadas()}}{
grafica de barras de secciones ganadas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$secciones_ganadas(
  bd = self$tab$aux$shp_secc,
  eleccion,
  eje_x = "",
  eje_y = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bd}}{es el shape de secciones luego de usar la función  `Tablero$filtrar`}

\item{\code{eleccion}}{elección de interés}

\item{\code{eje_x}}{leyenda del eje x}

\item{\code{eje_y}}{leyenda del eje y}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
objeto tipo ggplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-voto_relativo"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-voto_relativo}{}}}
\subsection{Method \code{voto_relativo()}}{
gráfica de barras con el total de votos obtenidos en términos absolutos y relativos
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$voto_relativo(
  partidos = self$tab$info$partidos,
  eleccion,
  eje_x = "",
  eje_y = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{partidos}}{heredado de la clase Tablero, partidos definidos desde la clase Electoral}

\item{\code{eleccion}}{elección de interés, debe estar en la clase Electoral}

\item{\code{eje_x}}{leyenda del eje x}

\item{\code{eje_y}}{leyenda del eje y}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
objeto de ggplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-distribucion_participacion"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-distribucion_participacion}{}}}
\subsection{Method \code{distribucion_participacion()}}{
gráfica de tipo violín con la distribución de la participación
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$distribucion_participacion(eleccion, eje_x = "", eje_y = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eleccion}}{elección de interés, debe estar en la clase Electoral}

\item{\code{eje_x}}{leyenda del eje x}

\item{\code{eje_y}}{leyenda del eje y
Gráfica de sankey}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-sankey"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-sankey}{}}}
\subsection{Method \code{sankey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$sankey()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
objeto tipo ggplot
Gráfica tipo pointrange
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-pointrange"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-pointrange}{}}}
\subsection{Method \code{pointrange()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$pointrange(indice, variables)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indice}}{ínndice de interés}

\item{\code{variables}}{determina si la gráfica va a ser de partidos o de elecciones, solo puede tomar "partidos" o "elecciones" como varlor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
objeto de ggplot
gráfica de mosaico
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-tiles"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-tiles}{}}}
\subsection{Method \code{tiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$tiles(indice, low = "#118ab2", high = "#ef476f")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indice}}{indice a comparar con el de participación}

\item{\code{low}}{color mínimo de la cloropleta}

\item{\code{high}}{color máximo de la cloropleta}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
objeto de ggplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
