% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.R
\name{Tablero}
\alias{Tablero}
\title{Clase R6 para leer y unir shapefiles}
\description{
Se inicia leyendo un shp
}
\details{
Al shp leído se le pude agregar otrabase de datos
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tablero-new}{\code{Tablero$new()}}
\item \href{#method-Tablero-agregar_eleccion}{\code{Tablero$agregar_eleccion()}}
\item \href{#method-Tablero-reiniciar_info}{\code{Tablero$reiniciar_info()}}
\item \href{#method-Tablero-clone}{\code{Tablero$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tablero-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tablero-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tablero$new(info_seccion)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tablero-agregar_eleccion"></a>}}
\if{latex}{\out{\hypertarget{method-Tablero-agregar_eleccion}{}}}
\subsection{Method \code{agregar_eleccion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tablero$agregar_eleccion(elecciones, nivel, bd_relacion, shp)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tablero-reiniciar_info"></a>}}
\if{latex}{\out{\hypertarget{method-Tablero-reiniciar_info}{}}}
\subsection{Method \code{reiniciar_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tablero$reiniciar_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tablero-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tablero-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tablero$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
